Carry 00128350's double through the land worker; record FLOOR's blockers
em_player_fall: 0017C580 passes 00128350's whole 64-bit $v0 (the double of
+220) unchanged as 001000E0's $a0; the land worker slots were int and
dropped the high word (60.0 is 0x404E000000000000). They are uint64_t now,
matching em_sdk_soft_float_00128350 and em_rvr_001000E0, and the oracle
compares whole $a0/$a1 (a low-word mutant fails).

FLOOR stays gated. FIRST_CONTROL.md "Missing today" inventories the ~450
closure worker slots; the blockers are the move walkers' 0019CB60 / 001A6440
(the slide and some fall edge tests reach them), 001755B0, record-level
climb helpers and the effect owner, being translated now.

+++ src/game/em_player_fall.h
@@ -122,10 +122,13 @@ typedef struct EmPlayerLandWorkers {
     int (*react_0021C120)(void *context, EmPlayerLiveActor *actor);
     int (*react_0021C350)(void *context, EmPlayerLiveActor *actor);
     int (*react_0021C270)(void *context, EmPlayerLiveActor *actor);
-    /* 00128350(+220): *result is its int return. 001000E0(a, b): *result is
-     * its return. */
-    int (*convert_00128350)(void *context, uint32_t value, int *result);
-    int (*test_001000E0)(void *context, int a, int b, int *result);
+    /* 00128350(+220): the float as a double, returned in the whole 64-bit
+     * $v0 (raw bits in *result). 0017C580 hands that register on unchanged
+     * as 001000E0's $a0, with $a1 = 0: 001000E0(a, b) compares the two
+     * doubles (*result is its return). em_sdk_soft_float_00128350 and
+     * em_rvr_001000E0 have these shapes. */
+    int (*convert_00128350)(void *context, uint32_t value, uint64_t *result);
+    int (*test_001000E0)(void *context, uint64_t a, uint64_t b, int *result);
     /* D_008106F1, read by 0017C580 after 00174AC0. */
     int (*progress_8106F1)(void *context, uint8_t *value);
 } EmPlayerLandWorkers;
